allow decimal image duration (#1619)
* add missing mysql migration

* allow decimal image duration

+++ ErsatzTV.Application/Images/Commands/UpdateImageFolderDuration.cs
@@ -1,3 +1,3 @@
 namespace ErsatzTV.Application.Images;
 
-public record UpdateImageFolderDuration(int LibraryFolderId, int? ImageFolderDuration) : IRequest<int?>;
+public record UpdateImageFolderDuration(int LibraryFolderId, double? ImageFolderDuration) : IRequest<double?>;
